[NFC] Refactor codegen patterns
Moves the CC patterns and Quake patterns into their own modules so they
can be reused rather than block copied.
