+++ include/cudaq/Optimizer/CodeGen/Passes.h
@@ -17,9 +17,13 @@
 #include "mlir/Pass/PassRegistry.h"
 
 namespace mlir {
+class LLVMTypeConverter;
 class MLIRContext;
 class Pass;
 class PassManager;
+namespace LLVM {
+class LLVMStructType;
+}
 } // namespace mlir
 
 namespace cudaq::opt {
@@ -51,6 +55,13 @@ void registerTargetPipelines();
 
 /// Register CodeGenDialect with the provided DialectRegistry.
 void registerCodeGenDialect(mlir::DialectRegistry &registry);
+
+mlir::LLVM::LLVMStructType lambdaAsPairOfPointers(mlir::MLIRContext *context);
+void populateCCToLLVMPatterns(mlir::LLVMTypeConverter &typeConverter,
+                              mlir::RewritePatternSet &patterns);
+void populateQuakeToLLVMPatterns(mlir::LLVMTypeConverter &typeConverter,
+                                 mlir::RewritePatternSet &patterns,
+                                 unsigned &measureCounter);
 
 // declarative passes
 #define GEN_PASS_DECL
